pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "esn-intraday"
version = "0.1.0"
description = "Multi-Horizon Echo State Network for Intraday Stock Return Prediction"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.10",
    "scikit-learn>=1.3",
    "pyyaml>=6.0",
    "matplotlib>=3.7",
    "tqdm>=4.65",
    "pyarrow>=14.0",
    "requests>=2.31",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
